Pass tests and copy Wan's comment changes

+++ test/execution/bandit_test.cpp
@@ -137,7 +137,7 @@ TEST_P(BanditTest, DISABLED_SimpleTest) {
 
   auto bandit = bandit::MultiArmedBandit(module.get(), action_names);
 
-  u32 num_actions = action_names.size();
+  u32 num_actions = static_cast<u32>(action_names.size());
   u32 num_trials = 200;
 
   std::vector<double> exec_time_individual(num_actions, 0.0);
