+++ test/execution/join_hash_table_test.cpp
@@ -211,7 +211,7 @@ TEST_F(JoinHashTableTest, DISABLED_PerfTest) {
     timer.Stop();
 
     auto mtps = (num_tuples / timer.elapsed()) / 1000.0;
-    auto size_in_kb = (concise ? ConciseTableFor(&join_hash_table)->GetTotalMemoryUsage()
+    auto size_in_kb = static_cast<double>(concise ? ConciseTableFor(&join_hash_table)->GetTotalMemoryUsage()
                                : GenericTableFor(&join_hash_table)->GetTotalMemoryUsage()) /
                       1024.0;
     EXECUTION_LOG_INFO("========== {} ==========", concise ? "Concise" : "Generic");
